:recycle: refactor: template 폴더 삭제 및 template 파일의 위치를 container로 변경
